+++ src/client/java/de/ottoextra/config/settings/OttoExtraSettingsScreen.java
@@ -92,8 +92,10 @@ public final class OttoExtraSettingsScreen extends Screen {
                 () -> open(new de.ottoextra.letter.ui.LetterEditorScreen(this, config)),
                 () -> open(new de.ottoextra.config.RegionThemeScreen(this, config)),
                 () -> open(new de.ottoextra.config.FollowingScreen(this, config)));
+        de.ottoextra.addon.OttoExtraAddons.registerSettings(this.registry);
         SettingsRegistry defaultRegistry = SettingsRegistry.build(new OttoExtraConfig(),
                 () -> { }, () -> { }, () -> { }, () -> { }, () -> { }, () -> { });
+        de.ottoextra.addon.OttoExtraAddons.registerDefaultSettings(defaultRegistry);
         for (SettingsRegistry.ModulePage m : defaultRegistry.modules()) {
             for (SettingsRegistry.Tab t : m.tabs()) {
                 for (SettingsRegistry.Card c : t.cards()) {